Leaving out constructor and destructor functions.
In the interest of cleanliness, simplicity, and/or ease of use, I think we
should leave these out of the library.

Revert "Add ctor/dtor capability.  I am semi-unhappy with the direct struct manipulation but wasn't happier with explicit get/set functions.  Furthermore, the semantics of the dtor call in the free function don't thrill me."

This reverts commit 518c870a7238c8d48a8881e49d9841d02d3804dd.